SWMAAS-1443 Create and utilize package directories
