--- Samples/kotlin/src/main/java/com/phunware/kotlin/sample/DemoAdapter.kt
@@ -1,73 +1,0 @@
-package com.phunware.kotlin.sample
-
-/* Copyright (C) 2018 Phunware, Inc.
-
-Permission is hereby granted, free of charge, to any person obtaining
-a copy of this software and associated documentation files (the
-"Software"), to deal in the Software without restriction, including
-without limitation the rights to use, copy, modify, merge, publish,
-distribute, sublicense, and/or sell copies of the Software, and to
-permit persons to whom the Software is furnished to do so, subject to
-the following conditions:
-
-The above copyright notice and this permission notice shall be
-included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-NONINFRINGEMENT. IN NO EVENT SHALL Phunware, Inc. BE LIABLE FOR ANY
-CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-Except as contained in this notice, the name of Phunware, Inc. shall
-not be used in advertising or otherwise to promote the sale, use or
-other dealings in this Software without prior written authorization
-from Phunware, Inc. */
-
-import android.support.v7.widget.RecyclerView
-import android.text.TextUtils
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-
-internal class DemoAdapter(private val demos: List<Demo>, private val listener: DemoOnClickListener) : RecyclerView.Adapter<DemoAdapter.ViewHolder>() {
-
-    internal interface DemoOnClickListener {
-        fun onItemClicked(title: String)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val v = LayoutInflater.from(parent.context).inflate(R.layout.demo_row, parent, false)
-        return ViewHolder(v, listener)
-    }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val d = demos[position]
-        holder.title.text = d.title
-        holder.description.text = d.description
-    }
-
-    override fun getItemCount(): Int = demos.size
-
-    fun getItem(title: String): Demo? {
-        if (!TextUtils.isEmpty(title)) {
-            demos
-                    .filter { it.title == title }
-                    .forEach { return it }
-        }
-        return null
-    }
-
-    internal class ViewHolder(itemView: View, private var listener: DemoOnClickListener)
-        : RecyclerView.ViewHolder(itemView) {
-        var title: TextView = itemView.findViewById(R.id.title)
-        var description: TextView = itemView.findViewById(R.id.description)
-
-        init {
-            itemView.setOnClickListener { listener.onItemClicked(title.text.toString()) }
-        }
-    }
-}
