--- Samples/kotlin/src/main/java/com/phunware/kotlin/sample/MarkerAnimation.kt
@@ -1,49 +1,0 @@
-package com.phunware.kotlin.sample
-
-/* Copyright (C) 2018 Phunware, Inc.
-
-Permission is hereby granted, free of charge, to any person obtaining
-a copy of this software and associated documentation files (the
-"Software"), to deal in the Software without restriction, including
-without limitation the rights to use, copy, modify, merge, publish,
-distribute, sublicense, and/or sell copies of the Software, and to
-permit persons to whom the Software is furnished to do so, subject to
-the following conditions:
-
-The above copyright notice and this permission notice shall be
-included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-NONINFRINGEMENT. IN NO EVENT SHALL Phunware, Inc. BE LIABLE FOR ANY
-CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-Except as contained in this notice, the name of Phunware, Inc. shall
-not be used in advertising or otherwise to promote the sale, use or
-other dealings in this Software without prior written authorization
-from Phunware, Inc. */
-
-import android.animation.ObjectAnimator
-import android.animation.TypeEvaluator
-import android.annotation.TargetApi
-import android.os.Build
-import android.util.Property
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.Marker
-
-internal object MarkerAnimation {
-
-    @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
-    fun animateMarkerTo(marker: Marker,
-                        finalPosition: LatLng,
-                        latLngInterpolator: LatLngInterpolator) {
-        val typeEvaluator = TypeEvaluator<LatLng> { fraction, startValue, endValue -> latLngInterpolator.interpolate(fraction, startValue, endValue) }
-        val property = Property.of(Marker::class.java, LatLng::class.java, "position")
-        val animator = ObjectAnimator.ofObject(marker, property, typeEvaluator, finalPosition)
-        animator.duration = 500
-        animator.start()
-    }
-}
