--- Samples/kotlin/src/main/java/com/phunware/kotlin/sample/PersonMarker.kt
@@ -1,55 +1,0 @@
-package com.phunware.kotlin.sample
-
-/* Copyright (C) 2018 Phunware, Inc.
-
-Permission is hereby granted, free of charge, to any person obtaining
-a copy of this software and associated documentation files (the
-"Software"), to deal in the Software without restriction, including
-without limitation the rights to use, copy, modify, merge, publish,
-distribute, sublicense, and/or sell copies of the Software, and to
-permit persons to whom the Software is furnished to do so, subject to
-the following conditions:
-
-The above copyright notice and this permission notice shall be
-included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-NONINFRINGEMENT. IN NO EVENT SHALL Phunware, Inc. BE LIABLE FOR ANY
-CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-Except as contained in this notice, the name of Phunware, Inc. shall
-not be used in advertising or otherwise to promote the sale, use or
-other dealings in this Software without prior written authorization
-from Phunware, Inc. */
-
-import com.google.android.gms.maps.model.Marker
-
-/**
- * A wrapper around a maker that associates a
- */
-class PersonMarker(val name: String?, val userType: String?, val marker: Marker?) {
-
-    override fun equals(o: Any?): Boolean {
-        if (this === o) return true
-        if (o == null || javaClass != o.javaClass) return false
-
-        val that = o as PersonMarker?
-
-        if (if (marker != null) marker != that!!.marker else that!!.marker != null) return false
-        if (if (name != null) name != that.name else that.name != null) return false
-        return if (userType != null) userType == that.userType else that.userType == null
-
-    }
-
-    override fun hashCode(): Int {
-        var result = marker?.hashCode() ?: 0
-        result = 31 * result + (name?.hashCode() ?: 0)
-        result = 31 * result + (userType?.hashCode() ?: 0)
-        return result
-    }
-}
-
