+++ Samples/kotlin/src/main/java/com/phunware/kotlin/sample/building/LoadBuildingActivity.kt
@@ -1,0 +1,120 @@
+package com.phunware.kotlin.sample.building
+
+/* Copyright (C) 2018 Phunware, Inc.
+
+Permission is hereby granted, free of charge, to any person obtaining
+a copy of this software and associated documentation files (the
+"Software"), to deal in the Software without restriction, including
+without limitation the rights to use, copy, modify, merge, publish,
+distribute, sublicense, and/or sell copies of the Software, and to
+permit persons to whom the Software is furnished to do so, subject to
+the following conditions:
+
+The above copyright notice and this permission notice shall be
+included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NONINFRINGEMENT. IN NO EVENT SHALL Phunware, Inc. BE LIABLE FOR ANY
+CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+Except as contained in this notice, the name of Phunware, Inc. shall
+not be used in advertising or otherwise to promote the sale, use or
+other dealings in this Software without prior written authorization
+from Phunware, Inc. */
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import android.util.Log
+import android.view.View
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
+import android.widget.Spinner
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.model.MapStyleOptions
+import com.phunware.core.PwCoreSession
+import com.phunware.kotlin.sample.R
+import com.phunware.kotlin.sample.building.adapter.FloorAdapter
+import com.phunware.mapping.MapFragment
+import com.phunware.mapping.OnPhunwareMapReadyCallback
+import com.phunware.mapping.PhunwareMap
+import com.phunware.mapping.manager.Callback
+import com.phunware.mapping.manager.PhunwareMapManager
+import com.phunware.mapping.model.Building
+import com.phunware.mapping.model.FloorOptions
+
+class LoadBuildingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback {
+    private lateinit var mapManager: PhunwareMapManager
+    private lateinit var currentBuilding: Building
+    private lateinit var spinnerAdapter: ArrayAdapter<FloorOptions>
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.load_building)
+
+        val floorSpinner = findViewById<Spinner>(R.id.floorSpinner)
+        spinnerAdapter = FloorAdapter(this)
+        floorSpinner.adapter = spinnerAdapter
+        floorSpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+            override fun onItemSelected(parent: AdapterView<*>, view: View, position: Int, id: Long) {
+                val floor = spinnerAdapter.getItem(id.toInt())
+                if (floor != null) {
+                    currentBuilding.selectFloor(floor.level)
+                }
+            }
+
+            override fun onNothingSelected(parent: AdapterView<*>) {}
+        }
+
+        // Create the map manager used to load the building
+        mapManager = PhunwareMapManager.create(this)
+
+        // Register the Phunware API keys
+        PwCoreSession.getInstance().registerKeys(this)
+
+        val mapFragment = fragmentManager.findFragmentById(R.id.map) as MapFragment
+        mapFragment.getPhunwareMapAsync(this)
+    }
+
+    override fun onPhunwareMapReady(phunwareMap: PhunwareMap) {
+        // Retrieve buildingId from integers.xml
+        val buildingId = resources.getInteger(R.integer.buildingId)
+
+        phunwareMap.googleMap.uiSettings.isMapToolbarEnabled = false
+        phunwareMap.googleMap.setMapStyle(MapStyleOptions.loadRawResourceStyle(
+                this@LoadBuildingActivity, R.raw.map_style))
+
+        mapManager.setPhunwareMap(phunwareMap)
+        mapManager.addBuilding(buildingId.toLong(),
+                object : Callback<Building> {
+                    override fun onSuccess(building: Building) {
+                        Log.d(TAG, "Building loaded successfully")
+                        currentBuilding = building
+
+                        // Populate floor spinner
+                        spinnerAdapter.clear()
+                        spinnerAdapter.addAll(building.buildingOptions.floors)
+
+                        // Set building to initial floor value
+                        val initialFloor = building.initialFloor()
+                        building.selectFloor(building.initialFloor().level)
+
+                        // Animate the camera to the building at an appropriate zoom level
+                        val cameraUpdate = CameraUpdateFactory
+                                .newLatLngBounds(initialFloor.bounds, 4)
+                        phunwareMap.googleMap.animateCamera(cameraUpdate)
+                    }
+
+                    override fun onFailure(throwable: Throwable) {
+                        Log.d(TAG, "Error when loading building -- " + throwable.message)
+                    }
+                })
+    }
+
+    companion object {
+        private val TAG = LoadBuildingActivity::class.java.simpleName
+    }
+}
