+++ Samples/kotlin/src/main/java/com/phunware/kotlin/sample/building/adapter/FloorAdapter.kt
@@ -1,0 +1,70 @@
+package com.phunware.kotlin.sample.building.adapter
+
+/* Copyright (C) 2018 Phunware, Inc.
+
+Permission is hereby granted, free of charge, to any person obtaining
+a copy of this software and associated documentation files (the
+"Software"), to deal in the Software without restriction, including
+without limitation the rights to use, copy, modify, merge, publish,
+distribute, sublicense, and/or sell copies of the Software, and to
+permit persons to whom the Software is furnished to do so, subject to
+the following conditions:
+
+The above copyright notice and this permission notice shall be
+included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NONINFRINGEMENT. IN NO EVENT SHALL Phunware, Inc. BE LIABLE FOR ANY
+CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+Except as contained in this notice, the name of Phunware, Inc. shall
+not be used in advertising or otherwise to promote the sale, use or
+other dealings in this Software without prior written authorization
+from Phunware, Inc. */
+
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ArrayAdapter
+import android.widget.TextView
+import com.phunware.kotlin.sample.R
+
+import com.phunware.mapping.model.FloorOptions
+
+import java.util.ArrayList
+
+internal class FloorAdapter(context: Context) : ArrayAdapter<FloorOptions>(context, 0, ArrayList()) {
+
+    override fun getView(position: Int, convertView: View?, parent: ViewGroup): View {
+        var v = convertView
+        val holder: ViewHolder
+        if (v == null) {
+            v = LayoutInflater.from(
+                    context).inflate(R.layout.floor_spinner_row, parent, false)
+            holder = ViewHolder()
+            holder.text = v!!.findViewById(android.R.id.text1)
+            v.tag = holder
+        } else {
+            holder = v.tag as ViewHolder
+        }
+        val floor = getItem(position)
+        if (floor != null) {
+            holder.text!!.text = floor.name
+        }
+
+        return v
+    }
+
+    override fun getDropDownView(position: Int, convertView: View?, parent: ViewGroup): View {
+        return getView(position, convertView, parent)
+    }
+
+    private class ViewHolder {
+        internal var text: TextView? = null
+    }
+}
