--- Samples/kotlin/src/main/java/com/phunware/kotlin/sample/utils/extensions/Context.kt
@@ -1,19 +1,0 @@
-package com.phunware.kotlin.sample.utils.extensions
-
-import android.content.Context
-import android.util.TypedValue
-import com.phunware.kotlin.sample.R
-
-/**
- * Returns the accent color as a ColorInt for the theme on the current context.
- */
-fun Context.accentColor(): Int {
-    val typedValue = TypedValue()
-
-    val a = obtainStyledAttributes(typedValue.data, intArrayOf(R.attr.colorAccent))
-    val color = a.getColor(0, 0)
-
-    a.recycle()
-
-    return color
-}
