+++ Samples/kotlin/src/main/res/layout/activity_routing.xml
@@ -21,7 +21,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <com.phunware.kotlin.sample.NavigationOverlayView
+        <com.phunware.kotlin.sample.routing.view.NavigationOverlayView
             android:id="@+id/nav_overlay"
             android:layout_width="match_parent"
             android:layout_height="93dp"
@@ -102,7 +102,7 @@
 
         <fragment
             android:id="@+id/routeSummaryFragment"
-            android:name="com.phunware.kotlin.sample.RouteSummaryFragment"
+            android:name="com.phunware.kotlin.sample.routing.fragment.RouteSummaryFragment"
             android:layout_width="0dp"
             android:layout_height="0dp"
             app:layout_constraintBottom_toTopOf="@+id/walk_time_view"
